# Faster parquet Dataset initialization on a local filesystem