This PR uses logic [from cudf](https://github.com/rapidsai/cudf/blob/c8e8526931443eb6121059c2a239d25b69741d37/python/dask_cudf/dask_cudf/io/parquet.py#L60) to better optimize the performance of `Dataset`.  It turns out that it is much faster to pass a file name than a file handle to `cudf.read_parquet` (as long as the filesystem is "local").

**Motivating Example (Criteo dataset)**:

```python
import nvtabular as nvt

#Before this PR
%time nvt.Dataset("/raid/dask-space/criteo/crit_pq_int", engine="parquet")

CPU times: user 856 ms, sys: 525 ms, total: 1.38 s
Wall time: 1.39 s

# After this PR
%time nvt.Dataset("/raid/dask-space/criteo/crit_pq_int", engine="parquet")

CPU times: user 847 ms, sys: 7.49 s, total: 8.34 s
Wall time: 8.34 s
```
<!--

Thank you for contributing to NVTabular :)

Here are some guidelines to help the review process go smoothly.

1. Please write a description in this text box of the changes that are being
   made.

2. Please ensure that you have written units tests for the changes made/features
   added.

3. If you are closing an issue please use one of the automatic closing words as
   noted here: https://help.github.com/articles/closing-issues-using-keywords/

4. If your pull request is not ready for review but you want to make use of the
   continuous integration testing facilities please label it with `[WIP]`.

5. If your pull request is ready to be reviewed without requiring additional
   work on top of it, then remove the `[WIP]` label (if present) and replace
   it with `[REVIEW]`. If assistance is required to complete the functionality,
   for example when the C/C++ code of a feature is complete but Python bindings
   are still required, then add the label `[HELP-REQ]` so that others can triage
   and assist. The additional changes then can be implemented on top of the
   same PR. If the assistance is done by members of the rapidsAI team, then no
   additional actions are required by the creator of the original PR for this,
   otherwise the original author of the PR needs to give permission to the
   person(s) assisting to commit to their personal fork of the project. If that
   doesn't happen then a new PR based on the code of the original PR can be
   opened by the person assisting, which then will be the PR that will be
   merged.

6. Once all work has been done and review has taken place please do not add
   features or make changes out of the scope of those requested by the reviewer
   (doing this just add delays as already reviewed code ends up having to be
   re-reviewed/it is hard to tell what is new etc!). Further, please do not
   rebase your branch on master/force push/rewrite history, doing any of these
   causes the context of any comments made by reviewers to be lost. If
   conflicts occur against master they should be resolved by merging master
   into the branch used for making the pull request.

Many thanks in advance for your cooperation!

-->